refactor(mobile): collapse ws-client web/native auth-recovery duplication
The web GraphQL-WS client copied ~150 lines of security-sensitive
credential-recovery and 4401→4403 remap logic from its native twin, so a
future fix to that path had to land twice.

Move the shared transport into `ws-client-core.ts` as `createWsClientModule`,
parameterised by the one genuine platform difference each fork owns: how the
underlying socket is built (React Native suppresses its Origin header; browsers
use the standard two-arg constructor) and which auth-store / auth-interceptor
fork supplies the credential helpers. `ws-client.ts` and `ws-client.web.ts`
become thin wrappers that wire those in.

Unify the re-declared rejection-result unions (NativeAuthRejectionResult /
WebAuthRejectionResult) into one shared `AuthRejectionResult` type both
interceptors import — the superset of both platforms' status labels, which the
shared retry check reads without knowing the platform.

No behavior change: the native and web ws-client test suites (52 tests) pass
unchanged, and each fork still constructs its own socket and imports its own
auth fork.
